Update the CountDownLatch API to use one type consistantly for a timeout value.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@920398 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/decaf/util/concurrent/CountDownLatch.h
@@ -107,7 +107,7 @@ namespace concurrent{
          * @throws InterruptedException - if the current thread is interrupted while waiting.
          * @throws Exception - if any other error occurs.
          */
-        virtual bool await( unsigned long timeOut )
+        virtual bool await( long long timeOut )
             throw ( decaf::lang::exceptions::InterruptedException,
                     decaf::lang::Exception );
